CleanCode for mspti.
# message auto-generated for no-merge-commit merge:
merge master into master

CleanCode for mspti

Created-by: msprofDeveloper

Match-id-244e26de31bb2813d514431eb33c541b609a7139

+++ mspti/activity/ascend/dev_prof_task.cpp
@@ -98,7 +98,7 @@ msptiResult DevProfTaskTsFw::StartTask()
     profStartPara.samplePeriod = SAMPLE_PERIOD;
     profStartPara.realTime = PROFILE_REAL_TIME;
     profStartPara.userData = &configP;
-    profStartPara.userDataSize = sizeof(TsTsFwProfileConfigT);
+    profStartPara.userDataSize = static_cast<unsigned int>(sizeof(TsTsFwProfileConfigT));
     int ret = ProfDrvStart(deviceId_, channelId_, &profStartPara);
     if (ret != 0) {
         MSPTI_LOGE("Failed to start TsTrackJob for device: %u, channel id: %u", deviceId_, channelId_);
@@ -136,7 +136,7 @@ msptiResult DevProfTaskStars::StartTask()
     profStartPara.samplePeriod = SAMPLE_PERIOD;
     profStartPara.realTime = PROFILE_REAL_TIME;
     profStartPara.userData = &configP;
-    profStartPara.userDataSize = sizeof(StarsSocLogConfigT);
+    profStartPara.userDataSize = static_cast<unsigned int>(sizeof(StarsSocLogConfigT));
     int ret = ProfDrvStart(deviceId_, channelId_, &profStartPara);
     if (ret != 0) {
         MSPTI_LOGE("Failed to start ProfStarsJob for device: %u, channel id: %u", deviceId_, channelId_);
